# server/public/api-docs/swagger.json
{
    "swagger": "2.0",
    "info": {
       "description": "A simple,user friendly Event Manager API for managing centers and hosting user events.",
       "version": "1.0.0",
       "title": "Event Manager.io",
       "contact": {
          "email": "iykay33@gmail.com"
       },
       "license": {
          "name": "Apache 2.0",
          "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
       }
    },
    "schemes": [
       "https",
       "http"
    ],
...
    "paths": {
...
       "/centers": {
          "post": {
             "tags": [
                "Center"
             ],
             "summary": "Add a new center to the store",
             "description": "adds a new center to the list of centers",
             "operationId": "addCenter",
             "consumes": [
                "application/x-www-form-urlencoded"
             ],
             "produces": [
                "application/json"
             ],
             "parameters": [
                {
                   "name": "Name",
                   "in": "formData",
                   "description": "The name of the center the user intends to add to the store.",
                   "required": true,
                   "type": "string"
                },
                {
                   "name": "image",
                   "in": "formData",
                   "description": "The image Url",
                   "type": "string"
                },
                {
                   "name": "description",
                   "in": "formData",
                   "description": "A short description of the center",
                   "required": true,
                   "type": "string",
                   "minLength": 10
                },
                {
                   "name": "capacity",
                   "in": "formData",
                   "description": "The capacity the center can hold",
                   "required": true,
                   "type": "number"
                },
                {
                   "name": "location",
                   "in": "formData",
                   "description": "The location of the center",
                   "required": true,
                   "type": "string"
                },
                {
                   "name": "price",
                   "in": "formData",
                   "description": "Price of the center",
                   "required": true,
                   "type": "number"
                }
             ],
             "responses": {
                "201": {
                   "description": "The new center was successfully created",
                   "schema": {
                      "type": "object",
                      "properties": {
                         "center": {
                            "type": "object",
                            "description": "The center details that was created "
                         },
                         "success": {
                            "type": "boolean",
                            "default": true,
                            "description": "And indication of if the operation was successful or not"
                         },
                         "message": {
                            "type": "string",
                            "description": "A message stating that the operation was successful"
                         }
                      }
                   }
                },
                "500": {
                   "description": "An error occured",
                   "schema": {
                      "$ref": "#/definitions/DefaultResponseStructure"
                   }
                }
             }
          },
          "get": {
             "tags": [
                "Center"
             ],
             "summary": "Returns all available center in the store",
             "description": "returns all available centers if no query parameter is present or limits the nmber of returned centers based on limit query params.it also searches center's store matching the name or category query params provided",
             "operationId": "getAllCenter",
             "consumes": [
                "application/xml"
             ],
             "produces": [
                "application/json"
             ],
             "parameters": [
                {
                   "name": "limit",
                   "in": "query",
                   "description": "max number of center to return",
                   "required": true,
                   "type": "integer",
                   "format": "int64"
                },
                {
                   "name": "name",
                   "in": "query",
                   "description": "name of center to search for",
                   "required": true,
                   "type": "integer",
                   "format": "int64"
                }
             ],
             "responses": {
                "200": {
                   "description": "The center was successfully retrieved",
                   "schema": {
                      "type": "object",
                      "properties": {
                         "center": {
                            "type": "object",
                            "description": "The center details that was created "
                         },
                         "success": {
                            "type": "boolean",
                            "default": true,
                            "description": "And indication of if the operation was successful or not"
                         },
                         "message": {
                            "type": "string",
                            "description": "A message stating that the operation was successful"
                         }
                      }
                   }
                },
                "404": {
                   "description": "The center was not found",
                   "schema": {
                      "$ref": "#/definitions/DefaultResponseStructure"
                   }
                },
                "500": {
                   "description": "An error occured",
                   "schema": {
                      "$ref": "#/definitions/DefaultResponseStructure"
                   }
                }
             }
          }
       },
       "/centers/centerId": {
          "put": {
             "tags": [
                "Center"
             ],
             "summary": "Updates a center with the given Id",
             "description": "",
             "operationId": "updateCenter",
             "consumes": [
                "application/x-www-form-urlencoded"
             ],
             "produces": [
                "application/json"
             ],
             "parameters": [
                {
                   "name": "centerId",
                   "in": "path",
                   "description": "ID of the center to be updated",
                   "required": true,
                   "type": "integer",
                   "format": "int64"
                },
                {
                   "name": "Name",
                   "in": "formData",
                   "description": "The name of the center the user intends to add to the store.",
                   "required": true,
                   "type": "string"
                },
                {
                   "name": "image",
                   "in": "formData",
                   "description": "The image Url",
                   "type": "string"
                },
                {
                   "name": "description",
                   "in": "formData",
                   "description": "A short description of the center",
                   "required": true,
                   "type": "string",
                   "minLength": 10
                },
                {
                   "name": "capacity",
                   "in": "formData",
                   "description": "The capacity the center can hold",
                   "required": true,
                   "type": "number"
                },
                {
                   "name": "location",
                   "in": "formData",
                   "description": "The location of the center",
                   "required": true,
                   "type": "string"
                },
                {
                   "name": "price",
                   "in": "formData",
                   "description": "Price of the center",
                   "required": true,
                   "type": "number"
                }
             ],
             "responses": {
                "201": {
                   "description": "The new center was successfully updated",
                   "schema": {
                      "type": "object",
                      "properties": {
                         "center": {
                            "type": "object",
                            "description": "The center details that was updated "
                         },
                         "success": {
                            "type": "boolean",
                            "default": true,
                            "description": "And indication of if the operation was successful or not"
                         },
                         "message": {
                            "type": "string",
                            "description": "A message stating that the operation was successful"
                         }
                      }
                   }
                },
                "404": {
                   "description": "The center was not found",
                   "schema": {
                      "$ref": "#/definitions/DefaultResponseStructure"
                   }
                },
                "500": {
                   "description": "An error occured",
                   "schema": {
                      "$ref": "#/definitions/DefaultResponseStructure"
                   }
                }
             }
          },
          "get": {
             "tags": [
                "Center"
             ],
             "summary": "Return a single center matching the Id provided",
             "description": "",
             "consumes": [
                "application/x-www-form-urlencoded"
             ],
             "produces": [
                "application/json"
             ],
             "parameters": [
                {
                   "name": "centerId",
                   "in": "path",
                   "description": "ID of center to return",
                   "required": true,
                   "type": "integer",
                   "format": "int64"
                }
             ],
             "responses": {
                "200": {
                   "description": "The operation was succesful",
                   "schema": {
                      "type": "object",
                      "properties": {
                         "center": {
                            "type": "object",
                            "description": "The center details that was retrieved "
                         },
                         "success": {
                            "type": "boolean",
                            "default": true,
                            "description": "And indication of if the operation was successful or not"
                         },
                         "message": {
                            "type": "string",
                            "description": "A message stating that the operation was successful"
                         }
                      }
                   }
                },
                "404": {
                   "description": "The center was not found",
                   "schema": {
                      "$ref": "#/definitions/DefaultResponseStructure"
                   }
                },
                "500": {
                   "description": "An error occured",
                   "schema": {
                      "$ref": "#/definitions/DefaultResponseStructure"
                   }
                }
             }
          }
       },
...
       "/centers/reviews/centerId": {
          "get": {
             "tags": [
                "Center"
             ],
             "summary": "gets all review for a center with the given Id",
             "description": "",
             "operationId": "getReviews",
             "consumes": [
                "application/x-www-form-urlencoded"
             ],
             "produces": [
                "application/json"
             ],
             "parameters": [
                {
                   "name": "centerId",
                   "in": "path",
                   "description": "ID of the center to be whose reviews are to be retrieved",
                   "required": true,
                   "type": "integer",
                   "format": "int64"
                }
             ],
             "responses": {
                "200": {
                   "description": "The operation was successful",
                   "schema": {
                      "type": "object",
                      "properties": {
                         "review": {
                            "type": "object",
                            "description": "The details of the review retrieved"
                         },
                         "success": {
                            "type": "boolean",
                            "default": true,
                            "description": "And indication of if the operation was successful or not"
                         },
                         "message": {
                            "type": "string",
                            "description": "A message stating that the operation was successful"
                         }
                      }
                   }
                },
                "500": {
                   "description": "An error occured",
                   "schema": {
                      "$ref": "#/definitions/DefaultResponseStructure"
                   }
                }
             }
          }
       },
...
    },
    "definitions": {
       "DefaultResponseStructure": {
          "type": "object",
          "properties": {
             "success": {
                "type": "boolean",
                "example": false,
                "description": "And indication of if the operation was successful or not"
             },
             "message": {
                "type": "string",
                "description": "An message sent by the api describing the result"
             }
          }
       },
...
    },
    "host": "virtserver.swaggerhub.com",
    "basePath": "/and84/ikenna/1.0.0"
 }